requirements.txt
streamlit
transformers
huggingface_hub
python-dotenv
torch